transcription/requirements.txt
faster-whisper
pyannote.audio
torch
